dRICH: test small sensor overlap
Do not merge.

This is a hacked version of the dRICH that only includes two sensors,
which have an obvious overlap in `jsroot`. In `athena` CI overlap
checks, this was missed, whereas the ECCE's `fun4all` overlap check did
catch this (they imported our geometry via GDML).

See https://eicweb.phy.anl.gov/EIC/detectors/athena/-/issues/151 for
more details.

The goal of this test is to see if the latest CI workflow still misses
this overlap.