test: post commit status with eicweb pipeline
This replaces the posting of a commit comment with updating a commit status. This status is attached to the pull_request, with a link to the eicweb pipeline.

GitHub actions tokens expire, so we cannot pass them to the triggered pipelines. We will have to figure out a way to pass the success of those pipelines back using a token in those pipelines.

For now we will need to check the pending status by hand and override.