feat: add header to compact file with liaisons
At this point we do not have a way to make it apparent who is the POC for each detector system. This adds an `info` header block to xml files with the `author`'s github handle for use in pull request review requests and assignment.

Note, `author` is the prescribed attribute in the [XML schema definition](https://www.lcsim.org/schemas/compact/1.0/compact.xsd) but it is intended to be interpreted as liaision. Multiple values for `author` can be added as a comma-separated list, so they can all be requested for review.