fix: calculate MGPD stave width based on envelope rmin, not mid-module rmin (#121)
### Briefly, what does this PR introduce?
This addresses the small overlaps between staves that have been bugging
the geant4 overlap checks. We were calculating the width of
perfectly-circle-filling staves by using the radius of the center of the
middle of the module, but then we were placing them starting at an
envelope rmin which was half a thickness smaller. This caused overlaps
at the inner corners for both inner and outer MPGD layers.

This change should make the staves a tad bit smaller (ratio of thickness
over radius of placement), cause a slightly smaller efficiency, but
should improve the pipeline stabilities by not triggering overlaps in
the MPGDs.

@mposik1983 I trust this is ok at least as a stopgap while better
solutions are developed.

### What kind of change does this PR introduce?
- [X] Bug fix (issue #__ unreported)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Slightly narrower MPGD staves.