feat: run epic_sciglass and _imaging benchmarks (#115)
### Briefly, what does this PR introduce?
In the past (ATHENA, ECCE) we were running benchmarks for multiple
configurations. This brings back that functionality for EPIC on GitHub.
At this point, the matrix is just epic_sciglass and epic_imaging, as the
likely two configurations in the October 2022 campaign.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #114)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
Multiple detector configurations will be tested by default, including
barrel imaging calorimeter, which has not been included in tests for a
while.