fix: revert to a single Hits collection for tracker (#52)
We currently write to a separate Hits collection for each barrel and
endcap tracker detector definition. This is required by some of the
dependencies because the collections are not read correctly if we don't
do this.

This branch moves everything back into a single collection to debug the
underlying issues and figure out where they exactly are coming from.

### Briefly, what does this PR introduce?
About half a year ago (or longer), when switching fully from dd4pod to
edm4hep, we had to put all hits in a dedicated collection instead of
reusing the collections across subdetectors. This reverts the behavior
as DD4hep versions 1.21 and later have fixed this issues.

### What kind of change does this PR introduce?
- [X] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
  - Benchmarks need updating:
- [x]
https://eicweb.phy.anl.gov/EIC/benchmarks/detector_benchmarks/-/merge_requests/130
- [x]
https://eicweb.phy.anl.gov/EIC/benchmarks/reconstruction_benchmarks/-/merge_requests/269
- [x]
https://eicweb.phy.anl.gov/EIC/benchmarks/physics_benchmarks/-/merge_requests/179
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators:
  - [x] Post on mattermost about this
  - [x] Let @eic/epic-devs know about this, in particular @mposik1983 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes. This now requires DD4hep 1.21 or newer for correct behavior (more
recent than LCG 101).

### Does this PR change default behavior?
Yes. By default the hits collections for vertex trackers, silicon barrel
and disk trackers, and MPGD trackers will not be numbered anymore.