feat: Remove fake DIRC until real DIRC is ready (#129)
### Briefly, what does this PR introduce?
This removes the current 'fake' DIRC implementation. It is not useful to
update the 'fake' DIRC in order to fix overlaps with the MPGDs due to
lack of correct DIRC geometry. We just remove the DIRC for now and aim
to include 'real' DIRC implementation soon.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #128)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators @mposik1983 @rdom 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No breaking changes, but will temporarily lead to lower material budget
in front of ECALs. @veprbl @mariakzurek

### Does this PR change default behavior?
No more DIRC. Less material budget before the ECAL.

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>