fix: decrease negative eta of cutout sciglass envelope
The change from -1.9 to -1.87 does not introduce overlaps in the sciglass and removes overlaps between the DIRC and the sciglass. With an MCP height of 23.746 cm starting at 70 cm radius, the DIRC is well within the 100 cm outer radius, so I think modifying the sciglass is the right approach here.