fix: decrease negative eta of cutout sciglass envelope (#236)
### Briefly, what does this PR introduce?
The DIRC and Ecal barrel sciglass envelope volume were overlapping about
2 cm at the DIRC readout box. This change from -1.9 to -1.87 does not
introduce overlaps in the sciglass and removes overlaps between the DIRC
and the sciglass. With an MCP height of 23.746 cm starting at 70 cm
radius, the DIRC is well within the 100 cm outer radius, so I think
modifying the sciglass is the right approach here.

### What kind of change does this PR introduce?
- [X] Bug fix (issue overlap fixed)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added: DIRC enabled in default
`epic.xml`
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators @veprbl 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.