feat: remove mpgd_dirc from brycecanyon (rework tracker config) (#242)
### Briefly, what does this PR introduce?
Until now, due to inaccuracy in the detector parameter table, the DIRC
MPGD was still included in the `brycecanyon` configuration. Taking the
`mpgd_dirc.xml` file out of the `tracker.xml` for only one configuration
was a bit cumbersome. I rewrote this to be a bit more flexible (and also
getting closer to the goal of interpreting the configuration yaml files
as a directory structure).

### What kind of change does this PR introduce?
- [X] Bug fix (issue: DIRC MPGD should have never been included in
`brycecanyon`)
- [X] New feature (issue: tracker configuration is too rigid in jinja
template)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators @mposik1983 @sly2j

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes. DIRC MPGD disappears from `brycecanyon`.

### Does this PR change default behavior?
Yes. DIRC MPGD disappears from `brycecanyon`.