feat: define default features inside jinja template, and assume features are file names (#244)
### Briefly, what does this PR introduce?
Having to allow for a non-existing `features` variable inside the jinja
template has been an annoying headache. It means that the default
configuration (that gives us `epic.xml`) is defined in an entirely
different way than the configuration files. This PR introduces a single
`features` definition at the top of the template for when `features` is
not defined (it is still slightly different because jinja is not yaml,
but it is much closer). The rest of the template can just assume that
`features` exists and legitimately crash when it doesn't.

The other aspect that this introduces is looping over the entries in
each detector type to load them. Rather than checking whether the `pid`
includes `drich` and then including `drich.xml`, we now just load
everything in the `pid` features, directly, interpreted as an xml file.
If you create a working prototype `trich.xml` in `compact/pid/` (with
triple radiators), you can simply add `trich` to the configurations.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #138)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No. This should result in the exact same xml files as before.

### Does this PR change default behavior?
No. This should result in the exact same xml files as before.

TODO
- [ ] double check when all commits in this branch are final

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>