starting point with HCAL sector steel (#117)
### Briefly, what does this PR introduce?
This PR is the start of introducing the correct oHCAL geometry for EPIC
associated with re-use of the sPHENIX outer hadronic calorimeter. At the
present time it just includes the oHCAL sector steel, including both the
normal HCAL sectors and the three modified "chimney" sectors that have a
cutout to accommodate the solenoid cryogenic stack.

While the sPHENIX simulations use a very complicated, full geometry for
the oHCAL based on a gdml export of the CAD files for the as-built
sectors, the geometry here is drastically simplified to include only the
longitudinal steel plates. The endplates, sector combs, and endplates
have been removed as these are small perturbations that are unlikely to
affect the overall performance.

I have, however, decided to continue with a gdml export for the
simplified geometry. The reason for this is that at the present time the
simplified geometry has a very simple mesh, and it is straightforward to
implement in a generic way in the xml using a TesslelatedSolid. The
geometry is created very quickly and the root viewer seems to have no
problem with the exported geometry. This approach allows us to add
additional relaism to the HCAL sectors by just updating the xml (gdml)
for the solid with no additional code changes. For example, the combs or
endplates with their curved surfaces would be added in trivially if it
was though to be necessary/interesting in the future.

The geometry is generated by importing the STEP file from the CAD design
and using the GDML wokrbench in FreeCAD to mesh the solid and write out
the gdml. The workbench has additional tools, like the ability to
decimate and/or simplify the mesh if desired. The resulting gdml is
added into the xml description for the detector.

The work that remains for the geometry is to: 

- Add in the scintillation tile shapes. There are 12 shapes, which an
additional special set of shapes in the chimney sector.
- Modify the stepping action to include the measured light yield maps.
This is not strictly necessary, but I would like to do it now so it
doesn't get forgotten in the future (or the information lost)

After this, there will need to be some work at the reconstruction level
to collect the scintillating tiles in groups of five corresponding to
the preamps on the calorimeter towers.

Here's a viewer image showing the beam pipe, solenoid and the new oHCAL.
The cryo cutout is visible at the top left.



![example](https://user-images.githubusercontent.com/3042746/190826365-41339269-afef-462c-aecd-2984aefdc256.png)


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ X] New feature (issue #110)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
At the present time this should be independent of the rest of the
detector geometry as the new outer HCAL geometry is stored in a new xml
file, and a new type "epic_oHcalBarrel" was created for this
development. This will aloo current simulations to continue until we are
ready to switch geometry.

### Does this PR change default behavior?
Yes - it replaces the existing ePIC HCAL geometry witha more realistic
version based on the sPHENIX HCAL.

Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>