fix: split overlap check jobs for barrel hcal (#261)
### Briefly, what does this PR introduce?
With the merge of the barrel hcal, we have temprorarily failing overlap
checks due to (it appears) a bug in the underlying classes. While this
is being solved, we want to continue having a working CI system, so we
disable the barrel hcal from some jobs. This PR moves stuff around so at
least the eicweb benchmarks pipelines are started again (even if they
may fail).

### What kind of change does this PR introduce?
- [X] Bug fix (issue: eicweb benchmark pipelines are not started)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.