43 rework backwards geometry for flexible studies (#50)
### Briefly, what does this PR introduce?
Complete reworking of the far backwards geometry with variable
positioning of tagger stations possible inside and outside of the main
beam vacuum. Constructed from boolean geometries to allow the extended
beamline vacuum and box to fit together correctly. Tagger stations
defined in separate xml to be constructed from additional technology
options added at a later date.

Documentation/comments will need to be added to explain how to adjust
the geometry.


### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #43 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Changes the readout structure for the backwards taggers, this might need
to be changed again.
The Lumi system has for now been commented out in the xml file until it
has been reworked on.

### Does this PR change default behavior?