fix: define Birks constants in materials definition (#307)
### Briefly, what does this PR introduce?
Until now we were relying on a local
[patch](https://github.com/eic/eic-spack/blob/v0.18/packages/geant4/birks.patch)
to geant4 to define Birks constants. That's error-prone and makes the
geometry dependent on external patches in a non-obvious way. This
introduces the Birks constants for those materials that are used in our
geometry, set to the values in the patch that has been in use:
- PbWO4: 0.0333*mm/MeV
- PlasticScint: 0.126*mm/MeV
- Polystyrene: 0.126*mm/MeV 
- SciGlass: 0.0333*mm/MeV

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators: @veprbl
@mariakzurek @lkosarz @johnlajoie @johnny8266

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.