Modification of ZDC (ecce-style) placement.  (#316)
### Briefly, what does this PR introduce?
Modification of the ecce-style ZDC placement. 
NOTE: it will not affect the current default ZDC behavior (Athena
style).
Misplacement is fixed. 
ZDC is shifted to the place in the CAD figure. 

### What kind of change does this PR introduce?
- [x] Bug fix (issue #309)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No. Only if one uses ZDC.xml instead of ZDC_Athena.xml, the position of
the ZDC will be modified.

Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>
Co-authored-by: Po-Ju Lin <pjlin@dyparton.phys.sinica.edu.tw>