Uncomment settings for high-acceptance, comment out high-divergence. (#335)
### Briefly, what does this PR introduce?

This changes the location of portions of the Roman Pots planes to make
them "high acceptance". It should be noted that it is obvious (now) that
the arrangement of the planes needs to be re-shuffled in the next
campaign to optimized the 10 sigma distance. Additionally, need to add
functionality to adjust the 10sigma position at run-time. Will do later
with a different PR.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: Not a bug necessarily - just have everything set for
high-divergence, switching to HA.

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

This PR makes a minimal change to the compact file, and only adjusts
some constants for the placement of the detector planes.

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No.

### Does this PR change default behavior?

Yes, it will increase the overall acceptance of far-forward protons at
low-Pt (in the Roman Pots, nowhere else).

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>