Fix Barrel HCAL Volume ID's To Allow Proper Clustering (#367)
### Briefly, what does this PR introduce?

This PR updates the volume ID's for the various detector elements
(sector, tower, tile) and corresponding readout configuration. The
"barrel" identifier has been removed. The fixed volume ID's now allow
for proper clustering, which requires an additional fix to EICrecon
which will be committed by Derek Anderson.

As a side benefit, this also includes simplified GDML solids for the
tiles. This work was done last fall but never merged in the rush to the
first simulation campaign.

### What kind of change does this PR introduce?
- [X] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

With the old volume ID configuration and EICrecon configuration we were
actually merging all the tiles in a sector, not all the tiles in a
tower, to create the merged hits.

 
![hitvsclustnum varyinghitmergefields_baseline e10th70n10kpip
d16m2y2023_720](https://user-images.githubusercontent.com/3042746/219803192-3cb66144-fa6e-400f-8d8e-94d51db700bb.png)

This first plot shows the old configuration - notice that the number of
merged hits and number of clusters is very similar, and this persisted
across all energies. This was a symptom that the merging was too broad
and we only saw a difference when a cluster split across a sector
boundary.

![hitvsclustnum varyinghitmergefields_onlytilesinhitmerger
e10th70n10kpip
d16m2y2023_720](https://user-images.githubusercontent.com/3042746/219803410-26ec6801-e638-48ed-b9f8-accd84a8c960.png)

This plot shows the new configuration, with towers energies created from
merged tiles. As you can see there are many more towers with energy that
clusters (since clusters should be a collection of towers).

Many thanks to Dmitry Kalinkin for his help with getting the volume ID's
straightened out!

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>