feat: use asymmetry tracker region in secondary particle handler (#376)
### Briefly, what does this PR introduce?
Due to limitations in DD4hep we were using symmetric limits here, but
DD4hep 1.24 allows asymmetric regions. This defines the "tracker region"
which is the region where created secondaries will be stored in the
MCParticles collection. This setting is backwards compatible in that
`tracker_region_zmax` will be used as the symmetric limit for older
DD4hep versions.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Christopher Dilks <c-dilks@users.noreply.github.com>