feat: explicitly define a world volume and material (#377)
### Briefly, what does this PR introduce?
By defining our own world volume, we gain some flexibility in defining
differently shaped world volumes, e.g. a central detector hall with far
forward and far backward tunnels. It also may allow us to treat the
world as a geant4 region with association user limit setters.

@c-dilks This has required moving the `optical_materials.xml` inclusion
outside the `<includes>` tag. When the file is included inside the
`<includes>` tag, it requires that the world is created at that time
since it is treated as a new translation unit. When the file is included
with a simple `<include>` tag, it includes the contents in-place in the
current translation unit. Can you please double-check that the change to
the optical_materials include line does not affect optical properties,
in particular your optical surface definitions?

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x] Changes have been communicated to collaborators @c-dilks 

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.