ci: only run prim file, dawn view on main (#395)
### Briefly, what does this PR introduce?
We are not able to keep up with workflows at times. While it is useful
to make sure all jobs run, even on pull requests, it is more useful to
make sure the visualization jobs run on main. This restricts prim file
generation and dawn views to main only.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.