fix: disable optical photon generation in hpDIRC (#442)
### Briefly, what does this PR introduce?
The hpDIRC is very inefficient right, causing 20x slowdowns for DIS
datasets. Since there is no reconstruction yet for the hpDIRC, there is
no reason to simulate all these photons.

This PR reverts the optical materials to their regular counterparts.

@niwgit This follows in part from today's discussion. There is no
digitization or reconstruction of the hpDIRC, so in the short term there
is no real use for these optical photons. Before re-enabling the optical
photons by default we would probably need digitization and
reconstruction, and we need a better understanding of the impact on
running times.

Cc: @mdiefent @rahmans1 @T-Britton 

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- We need a benchmark to catch regressions in running time like this...
@veprbl @torrijeske
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes, reverts to not having optical photons in the hpDIRC.

### Does this PR change default behavior?
Yes, no more optical photons in the hpDIRC.