fix: dd4hep::CartesianField renamed type to field_type (#449)
### Briefly, what does this PR introduce?
In the next DD4hep version, the field type will be stored in a variable
`field_type` not `type`. This will cause epic to fail to compile. Of
course, until then this version will fail to compile. And all older
geometries will fail to compile with the newer DD4hep as well. So that's
kinda annoying and something I don't have a good solution to...

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.