ci: consolidate geoviewer job, include epic_ in detector_config json (#478)
### Briefly, what does this PR introduce?
We create a list of detector configs in an early step and use it later
on. This expands that functionality to create both a csv and json list.
Both have now `epic_<configuration>` as format, consistent with other
uses of `DETECTOR_CONFIG`.

This PR also avoids a large number of CI jobs on eicweb to trigger many
geoviewers: each CI job was building the entire geometry again. We can
avoid this, and just create the geoviewer artifacts in one go. We pass
the csv list for that reason.

We also turn the matrix jobs for fast operations back into sequential
jobs, avoiding the repeated overhead of starting a container
environment.

TODO:
- [x] merge
https://eicweb.phy.anl.gov/EIC/benchmarks/geoviewer/-/merge_requests/2
if success
- [x] modify back to using `main` instead of `detector-config-csv-list`

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>