Update ff magnet vacuum (#524)
### Briefly, what does this PR introduce?

This PR updates the magnet FF vacuum to remove gaps and overlaps and
properly use CutTubes to connect the inner pieces together.
Additionally, it has been modified to take the magnet locations as its
primary input, and calculate everything relative to that, which should
make maintenance easier in the long-run.

It should be noted that this volume is ONLY for the region between the
end of the central beam pipe and the end of the b1apf magnet. The region
occupied by the RP and OMD is handled separately.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x] Tests for the changes have been added
- [x] Documentation has been added / updated
- [x] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, it should further improve basic functionality.

### Does this PR change default behavior?

Yes, it make the vacuum volume in the FF region more-continuous.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>