Update solenoid material budget (#529)
### Briefly, what does this PR introduce?

This PR updates the ePIC Solenoid (MARCO) to the material specifications
provided by the design team located at:

https://wiki.bnl.gov/EPIC/index.php?title=Experimental_Solenoid

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: Update solenoid material map

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No breaking changes 

### Does this PR change default behavior?

No, it just updates the material budget for the MARCO solenoid.

This PR is a response to a material map run for me by Nicholas Schmidt
at ORNL - you can see below that the solenoid is almost an interaction
length, when it should be closer to 0.5 interaction lengths according to
the designers:


![image](https://github.com/eic/epic/assets/3042746/f724417b-7f82-4b4c-a29c-2819a1edf413)

The new material map reduces the material budget for the solenoid to be
a bit closer to what we would expect:


![image](https://github.com/eic/epic/assets/3042746/a9678b41-9147-458e-b938-dda924dac3c9)


![image](https://github.com/eic/epic/assets/3042746/90170d67-b8ab-43d5-a66e-b8497f5d3d5f)

Fully updated based on the material map from Renuka and Valerio dated 17
Sept 2023. The endcaps got a bit thicker, but most of the barrel is
essentially the same. I removed the He layers that were in the original
based on a comment from Valerio:

"The total volume of helium circulating inside the pipes is just 6000
mm3, corresponding to an uniform layer of Helium of 0.05 mm around the
magnet. We neglected it for this reason, but if you think it is
important we can add it."

[Material_in_the_magnet_Marco_2T_September.17.2023
(1).xlsx](https://github.com/eic/epic/files/12705405/Material_in_the_magnet_Marco_2T_September.17.2023.1.xlsx)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>