fix: ensure file download command is constexpr since required in C++20 (#520)
### Briefly, what does this PR introduce?
In C++20 (and/or fmt-10.1.0) we start running into `constexpr`
requirements for format strings. No problem, here we make them
`constexpr`. This removes the ability to set them from the compact
files, which no one was using.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #519)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
Yes, if you had a local compact file which set the `cmd:` value in the
bfield loader, you will have to make sure you install `curl` as that
will be used.

### Does this PR change default behavior?
No.

Co-authored-by: Chao Peng <pengc2010@gmail.com>