Zdc SiPM-on-tile (#534)
### Briefly, what does this PR introduce?
Adds an SiPM-on-tile implementation of the ZDC Hcal, using similar
technology to the forward calorimeter insert.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #530)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, it does not introduce breaking changes

### Does this PR change default behavior?

No, it does not change default behavior


To test this geometry,
Follow instructions in README.md to install epic and run source
install/setup.sh
Run ./run_sim_hepmc_zdc inside the attached test_suite.zip in order to
run a quick simulation of events in the proposed detector.

[test_suite.zip](https://github.com/eic/epic/files/12708021/test_suite.zip)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>